minimize the example demo
- improve README
- refactor the ScanView into packages for easier usage
